Enlarge the header to 100vh Add info card styling
Enlarged the header element to now take up the whole viewport, this was
done due to stylistic possibilities that such a layout might bring abt.

As for the card styling, all this did was give shape to the html ele-
ments. The basic skeleton of these cards are complete but the fleshing
out the fine details is still warranted.