Add info cards to the landing page
Adds 3 info cards (Section Elements) to the main section of the landing
page, that briefly describe the various services provided by the
company, with "read more" buttons at the bottom redirecting the user to
the respective resources for further elucidation. The section Elements
are subject to replacement if in future supplanting them with a
different element semantically makes sense.