Make navbar fixed Remove fixed from top-panel
Makes Navbar fixed instead of the previously top-panel. This was done
to make the logo stay within the confines of the header instead of
following the user as they scroll. Adds Sticky position to the logo so
that it scrolls along within header. Also adjusts the width for the
navbar and centers it with left: property and Transform: property;
Might be subject to change later on.