Change Navbar to display services separately
Changed the nav bar in index.html and styles.css to display the company
services separately instead of one overarching services page and
products page. Did this because currently the company does not possess
actual projects to display so it might be awkward. This will change as
the time passes.