Fix Navbar squishing on resize
Fixes the issue where the navbar would squish after it was resized to
below 800px viewport width. The problem was the top panel being set to
a child element height on resizing (Which on phone devices would
be 0). This update tells the site to set the top panel size back to its
defined, tallest element when switching back to the desktop view.