Make logo dynamically updated based on brightness
Adds a functionality to the site logo to change based on the average
brightness of the carousel's currently displayed picture. It calculates
the average brightness of all pixels in the area occupied by the logo.
If average is higher than threshold, the site loads in a different
picture for the logo and vise versa. The draw call and updating only
happens when a new slide is loaded in, in other words, only on carousel
updates.