Make the logo change colors on screen resize
Now adds a behavior that makes the logo change colors on resize change
if needed, because there were cases when after exiting mobile view, if
the slide was bright the logo did not change into dark mode. Now the
website checks on every resize (if the viewport size is sufficient)
if the logo needs mode change. Also disables the logo switcher beow 1220
screen width.